last

+++ kline/types.go
@@ -1,7 +1,7 @@
 package kline
 
 type Kline struct {
-	Close       float64
-	Volume      float64
-	TakerVolume float64
+	Close       float32
+	Volume      float32
+	TakerVolume float32
 }
