+++ model/dynparam/dynparam.go
@@ -3,5 +3,5 @@ package dynparam
 import "github.com/aszxqaz/model/kline"
 
 type DynamicParam interface {
-	Evaluate(klines []kline.Kline) (float64, bool)
+	Evaluate(klines []kline.Kline) (float32, bool)
 }
